chore: move IATP Transformer Extension into IATP module
