+++ extensions/common/iam/identity-trust/identity-trust-core/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -15,3 +15,4 @@
 org.eclipse.edc.iam.identitytrust.core.IatpDefaultServicesExtension
 org.eclipse.edc.iam.identitytrust.core.IatpScopeExtractorExtension
 org.eclipse.edc.iam.identitytrust.core.IdentityAndTrustExtension
+org.eclipse.edc.iam.identitytrust.core.IdentityTrustTransformExtension
