+++ extensions/common/iam/identity-trust/identity-trust-core/src/main/resources/document/credentials.v2.jsonld
