+++ extensions/common/iam/identity-trust/identity-trust-core/src/main/resources/document/iatp.v08.jsonld
